+++ resources/wx_gui/gui_main_menu.py
@@ -8,10 +8,11 @@ import webbrowser
 from resources.wx_gui import (
     gui_build,
     gui_macos_installer_download,
-    gui_sys_patch,
     gui_support,
     gui_help,
     gui_settings,
+    gui_sys_patch_start,
+    gui_sys_patch_display,
     gui_update,
 )
 from resources import (
@@ -24,6 +25,7 @@ from data import os_data
 
 class MainFrame(wx.Frame):
     def __init__(self, parent: wx.Frame, title: str, global_constants: constants.Constants, screen_location: tuple = None):
+        logging.info("Initializing Main Menu Frame")
         super(MainFrame, self).__init__(parent, title=title, size=(600, 400), style=wx.DEFAULT_FRAME_STYLE & ~(wx.RESIZE_BORDER | wx.MAXIMIZE_BOX))
         gui_support.GenerateMenubar(self, global_constants).generate()
 
@@ -39,6 +41,7 @@ class MainFrame(wx.Frame):
 
         self.Centre()
         self.Show()
+
 
         self._preflight_checks()
 
@@ -213,11 +216,11 @@ class MainFrame(wx.Frame):
             pop_up.ShowModal()
 
             if pop_up.GetReturnCode() != wx.ID_YES:
-                print("- Skipping OpenCore and root volume patch update...")
+                print("Skipping OpenCore and root volume patch update...")
                 return
 
 
-            print("- Updating OpenCore and root volume patches...")
+            print("Updating OpenCore and root volume patches...")
             self.constants.update_stage = gui_support.AutoUpdateStages.CHECKING
             self.Hide()
             pos = self.GetPosition()
@@ -277,7 +280,7 @@ class MainFrame(wx.Frame):
 
 
     def on_post_install_root_patch(self, event: wx.Event = None):
-        gui_sys_patch.SysPatchFrame(
+        gui_sys_patch_display.SysPatchDisplayFrame(
             parent=self,
             title=self.title,
             global_constants=self.constants,