+++ CHANGELOG.md
@@ -14,7 +14,20 @@
     - ex. OpenCore-Patcher.app
 - Backend changes:
   - Call `setpgrp()` to prevent app from being killed if parent process is killed (ie. LaunchAgents)
+  - Rework logging handler:
+    - Implement formatted logging
+      - Allowing easier debugging
+    - Implement per-version, per-run file logging
+      - ex. OpenCore-Patcher (0.6.7) (2021-12-31-12-34-56).log
+    - Keep only 10 latest log files
+    - Reveal log file in Finder on main thread crash
+    - Avoid writing username to log file
+  - Resolve SharedSupport.dmg pathing error during macOS Installer Verification
+    - Applicable to systems with 2 (or more) USB Installers with the same name plugged in
   - Resolve payloads path being mis-routed during CLI calls
+  - Add UI when fetching root patches for host
+  - Remove progress bar work-around for non-Metal in Monterey and later
+    - Requires host to have been patched with PatcherSupportPkg 1.1.2 or newer
 - Increment Binaries:
   - PatcherSupportPkg 1.1.2 - release